requirements.txt
Django==5.0.1
django-countries==7.5.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
gunicorn==21.2.0
langchain==0.1.0
langchain-community==0.0.13
langchain-core==0.1.10
langchain-openai==0.0.2.post1
openai==1.7.2
psycopg2_binary==2.9.9
pypdf==3.17.4
qdrant-client==1.7.0
django-cors-headers==4.3.1